requirements.txt
pillow
opencv-python
imagehash
unidecode